hackathon/mBrainAligner/src/3D_U-Net/requirements.txt
bleach==1.5.0
cycler==0.10.0
decorator==4.1.2
entrypoints==0.2.3
enum34==1.1.6
h5py==2.7.1
html5lib==0.9999999
ipykernel==4.6.1
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.3
jedi==0.11.0
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.1.2
keras-contrib==2.0.8
Markdown==2.6.9
MarkupSafe==1.0
matplotlib==2.1.0
mistune==0.8
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.2.1
numexpr==2.7.0
numpy==1.13.3
olefile==0.44
pandas==0.21.0
pandocfilters==1.4.2
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.3.0
prompt-toolkit==1.0.15
protobuf==3.4.0
ptyprocess==0.5.2
pycurl==7.43.0
Pygments==2.2.0
pygobject==3.20.0
pyparsing==2.2.0
python-apt==1.1.0b1
python-dateutil==2.6.1
pytz==2017.3
PyYAML==5.1.2
pyzmq==16.0.3
qtconsole==4.3.1
scikit-learn==0.19.1
scipy==1.0.0
simplegeneric==0.8.1
SimpleITK==1.2.3
six==1.11.0
sklearn==0.0
tables==3.4.0
tensorflow-gpu==1.4.0
tensorflow-tensorboard==0.4.0rc2
terminado==0.6
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
widgetsnbextension==3.0.6
